requirements.txt
pandas>=2.0
pyyaml>=6.0
matplotlib>=3.7
pytest>=7.0
